app/pyproject.toml
[project]
name = "base-iq-insights-dashboard"
version = "0.1.0"
description = "primary kpis"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "alphashape>=1.3.1",
    "base-core-lib",
    "bs4>=0.0.2",
    "celery>=5.5.3",
    "cryptography>=45.0.6",
    "dash>=3.2.0",
    "dash-bootstrap-components>=2.0.4",
    "duckdb>=1.3.2",
    "duckdb-engine>=0.17.0",
    "flask>=3.1.2",
    "flask-dance>=7.1.0",
    "geodatasets>=2024.8.0",
    "geopandas>=1.1.1",
    "google-cloud-secret-manager>=2.24.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
]

[tool.uv.sources]
base-core-lib = { git = "https://github.com/nearl-base/base-core-lib.git" }
